% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RemoteMatrix.R
\name{dim.RemoteMatrix}
\alias{dim.RemoteMatrix}
\title{The dim method for the S4 RemoteMatrix class.}
\usage{
\method{dim}{RemoteMatrix}(x)
}
\description{
The dim method for the S4 RemoteMatrix class.
}

