% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RemoteMatrix.R
\docType{methods}
\name{initialize,RemoteMatrix-method}
\alias{initialize,RemoteMatrix-method}
\title{The constructor for the S4 RemoteMatrix class.}
\usage{
\S4method{initialize}{RemoteMatrix}(.Object, uri)
}
\description{
The constructor for the S4 RemoteMatrix class.
}

