% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RemoteMatrix.R
\docType{class}
\name{RemoteMatrix-class}
\alias{RemoteMatrix-class}
\title{An S4 class to represent a remote matrix.}
\description{
An S4 class to represent a remote matrix.
}
\section{Slots}{

\describe{
\item{\code{uri}}{The URI to the remote matrix.}
}}

