% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RemoteMatrix.R
\docType{package}
\name{RemoteMatrix}
\alias{RemoteMatrix}
\alias{RemoteMatrix-package}
\title{RemoteMatrix: A package for remote matrices.}
\description{
RemoteMatrix: A package for remote matrices.
}

