% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RemoteMatrix.R
\docType{methods}
\name{[,RemoteMatrix,ANY,ANY-method}
\alias{[,RemoteMatrix,ANY,ANY-method}
\title{The subset operator for the S4 RemoteMatrix class.}
\usage{
\S4method{[}{RemoteMatrix,ANY,ANY}(x, i, j, drop)
}
\description{
The subset operator for the S4 RemoteMatrix class.
}

