% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RemoteMatrix.R
\docType{methods}
\name{[<-,RemoteMatrix-method}
\alias{[<-,RemoteMatrix-method}
\title{The replacement operator for the S4 RemoteMatrix class.}
\usage{
\S4method{[}{RemoteMatrix}(x, i, j) <- value
}
\description{
The replacement operator for the S4 RemoteMatrix class.
}

